Start of signal refactor
Split mqtt and notifications into separate imports in prep for our
notification rewrite.
