+++ pomodoro/apps.py
@@ -6,8 +6,11 @@ logger = logging.getLogger(__name__)
 
 
 class PomodoroConfig(AppConfig):
-    name = 'pomodoro'
+    name = "pomodoro"
 
     def ready(self):
-        import pomodoro.tasks  # NOQA
         import pomodoro.signals  # NOQA
+        import pomodoro.signals.notification  # NOQA
+
+        if self.apps.is_installed("dmqtt"):
+            import pomodoro.signals.mqtt  # NOQA
